+++ graphql-native/src/main/java/io/ballerina/stdlib/graphql/runtime/engine/ResponseGenerator.java
@@ -20,13 +20,11 @@ package io.ballerina.stdlib.graphql.runtime.engine;
 
 import io.ballerina.runtime.api.Environment;
 import io.ballerina.runtime.api.TypeTags;
-import io.ballerina.runtime.api.async.Callback;
 import io.ballerina.runtime.api.creators.TypeCreator;
 import io.ballerina.runtime.api.creators.ValueCreator;
 import io.ballerina.runtime.api.types.ArrayType;
 import io.ballerina.runtime.api.utils.StringUtils;
 import io.ballerina.runtime.api.values.BArray;
-import io.ballerina.runtime.api.values.BError;
 import io.ballerina.runtime.api.values.BMap;
 import io.ballerina.runtime.api.values.BObject;
 import io.ballerina.runtime.api.values.BString;
@@ -35,11 +33,9 @@ import io.ballerina.runtime.api.values.BValue;
 
 import java.util.ArrayList;
 import java.util.List;
-import java.util.concurrent.CountDownLatch;
 
 import static io.ballerina.stdlib.graphql.runtime.engine.Engine.executeResource;
 import static io.ballerina.stdlib.graphql.runtime.engine.Engine.getArgumentsFromField;
-import static io.ballerina.stdlib.graphql.runtime.engine.EngineUtils.ERRORS_FIELD;
 import static io.ballerina.stdlib.graphql.runtime.engine.EngineUtils.IS_FRAGMENT_FIELD;
 import static io.ballerina.stdlib.graphql.runtime.engine.EngineUtils.KEY;
 import static io.ballerina.stdlib.graphql.runtime.engine.EngineUtils.NAME_FIELD;
@@ -48,68 +44,43 @@ import static io.ballerina.stdlib.graphql.runtime.engine.EngineUtils.ON_TYPE_FIE
 import static io.ballerina.stdlib.graphql.runtime.engine.EngineUtils.SELECTIONS_FIELD;
 import static io.ballerina.stdlib.graphql.runtime.engine.EngineUtils.copyAndUpdateResourcePathsList;
 import static io.ballerina.stdlib.graphql.runtime.engine.EngineUtils.createDataRecord;
-import static io.ballerina.stdlib.graphql.runtime.engine.EngineUtils.getErrorDetailRecord;
 import static io.ballerina.stdlib.graphql.runtime.engine.EngineUtils.getNameFromRecordTypeMap;
 import static io.ballerina.stdlib.graphql.runtime.engine.EngineUtils.isScalarType;
 
 /**
- * This class is used as a callback class for Ballerina resource execution.
+ * Used to generate the response for a GraphQL request in Ballerina.
  */
-public class CallableUnitCallback implements Callback {
-    private final Environment environment;
-    private final CountDownLatch latch;
-    private final BObject visitor;
-    private final BObject node;
-    private final BMap<BString, Object> data;
-
-    public CallableUnitCallback(Environment environment, CountDownLatch latch, BObject visitor, BObject node,
-                                BMap<BString, Object> data) {
-        this.environment = environment;
-        this.latch = latch;
-        this.visitor = visitor;
-        this.node = node;
-        this.data = data;
+public class ResponseGenerator {
+    private ResponseGenerator() {
     }
 
-    @Override
-    public void notifySuccess(Object o) {
-        if (o instanceof BError) {
-            BError bError = (BError) o;
-            appendErrorToVisitor(bError);
-        } else {
-            getDataFromResult(this.environment, this.visitor, this.node, o, this.data);
-        }
-        this.latch.countDown();
-    }
-
-    @Override
-    public void notifyFailure(BError bError) {
-        appendErrorToVisitor(bError);
-        this.latch.countDown();
-    }
-
-    public static void getDataFromResult(Environment environment, BObject visitor, BObject node, Object result,
-                                         BMap<BString, Object> data) {
-        if (result instanceof BMap) {
-            BMap<BString, Object> map = (BMap<BString, Object>) result;
-            if (map.getType().getTag() == TypeTags.RECORD_TYPE_TAG) {
-                getDataFromRecord(environment, visitor, node, map, data);
-            } else if (map.getType().getTag() == TypeTags.MAP_TAG) {
-                getDataFromMap(environment, visitor, node, map, data);
-            }
-        } else if (result instanceof BArray) {
-            getDataFromArray(environment, visitor, node, (BArray) result, data);
-        } else if (result instanceof BTable) {
-            getDataFromTable(environment, visitor, node, (BTable) result, data);
-        } else if (result instanceof BObject) {
-            getDataFromService(environment, (BObject) result, visitor, node, data, new ArrayList<>());
+    static void getDataFromResult(Environment environment, BObject visitor, BObject node, Object result,
+                                  BMap<BString, Object> data, CallbackHandler callbackHandler) {
+        if (result instanceof BValue) {
+            int tag = ((BValue) result).getType().getTag();
+            if (tag < TypeTags.JSON_TAG) {
+                data.put(node.getStringValue(NAME_FIELD), result);
+            } else if (tag == TypeTags.RECORD_TYPE_TAG) {
+                getDataFromRecord(environment, visitor, node, (BMap<BString, Object>) result, data, callbackHandler);
+            } else if (tag == TypeTags.MAP_TAG) {
+                getDataFromMap(environment, visitor, node, (BMap<BString, Object>) result, data, callbackHandler);
+            } else if (tag == TypeTags.ARRAY_TAG) {
+                getDataFromArray(environment, visitor, node, (BArray) result, data, callbackHandler);
+            } else if (tag == TypeTags.TABLE_TAG) {
+                getDataFromTable(environment, visitor, node, (BTable) result, data, callbackHandler);
+            } else if (tag == TypeTags.SERVICE_TAG) {
+                getDataFromService(environment, (BObject) result, visitor, node, data, new ArrayList<>(),
+                                   callbackHandler);
+            } // Here, `else` should not be reached.
         } else {
             data.put(node.getStringValue(NAME_FIELD), result);
         }
     }
 
     static void getDataFromService(Environment environment, BObject service, BObject visitor, BObject node,
-                                   BMap<BString, Object> data, List<String> paths) {
+                                   BMap<BString, Object> data, List<String> paths, CallbackHandler handler) {
+        ResourceCallback callback = new ResourceCallback(environment, visitor, node, createDataRecord(), handler);
+        handler.addCallback(callback);
         BArray selections = node.getArrayValue(SELECTIONS_FIELD);
         BMap<BString, Object> subData = createDataRecord();
         for (int i = 0; i < selections.size(); i++) {
@@ -118,34 +89,18 @@ public class CallableUnitCallback implements Callback {
             BObject subNode = selection.getObjectValue(NODE_FIELD);
             if (isFragment) {
                 if (service.getType().getName().equals(subNode.getStringValue(ON_TYPE_FIELD).getValue())) {
-                    executeResourceForFragmentNodes(environment, service, visitor, subNode, subData, paths);
+                    executeResourceForFragmentNodes(environment, service, visitor, subNode, subData, paths, handler);
                 }
             } else {
-                List<String> updatedPaths = copyAndUpdateResourcePathsList(paths, subNode);
-                executeResource(environment, service, visitor, subNode, subData, updatedPaths);
+                executeResourceWithPath(environment, visitor, subNode, service, subData, paths, handler);
             }
             data.put(node.getStringValue(NAME_FIELD), subData);
         }
-    }
-
-    static void getDataFromArray(Environment environment, BObject visitor, BObject node, BArray result,
-                                 BMap<BString, Object> data) {
-        if (isScalarType(result.getElementType())) {
-            data.put(node.getStringValue(NAME_FIELD), result);
-        } else {
-            BArray resultArray = ValueCreator.createArrayValue(getDataRecordArrayType());
-            for (int i = 0; i < result.size(); i++) {
-                Object resultElement = result.get(i);
-                BMap<BString, Object> subData = createDataRecord();
-                getDataFromResult(environment, visitor, node, resultElement, subData);
-                resultArray.append(subData.get(node.getStringValue(NAME_FIELD)));
-            }
-            data.put(node.getStringValue(NAME_FIELD), resultArray);
-        }
+        handler.markComplete(callback);
     }
 
     static void getDataFromRecord(Environment environment, BObject visitor, BObject node, BMap<BString, Object> record,
-                                  BMap<BString, Object> data) {
+                                  BMap<BString, Object> data, CallbackHandler callbackHandler) {
         BArray selections = node.getArrayValue(SELECTIONS_FIELD);
         BMap<BString, Object> subData = createDataRecord();
         for (int i = 0; i < selections.size(); i++) {
@@ -154,72 +109,88 @@ public class CallableUnitCallback implements Callback {
             BObject subNode = selection.getObjectValue(NODE_FIELD);
             if (isFragment) {
                 if (subNode.getStringValue(ON_TYPE_FIELD).getValue().equals(getNameFromRecordTypeMap(record))) {
-                    processFragmentNodes(environment, visitor, subNode, record, subData);
+                    processFragmentNodes(environment, visitor, subNode, record, subData, callbackHandler);
                 }
             } else {
                 BString fieldName = subNode.getStringValue(NAME_FIELD);
                 Object fieldValue = record.get(fieldName);
-                getDataFromResult(environment, visitor, subNode, fieldValue, subData);
+                getDataFromResult(environment, visitor, subNode, fieldValue, subData, callbackHandler);
             }
         }
         data.put(node.getStringValue(NAME_FIELD), subData);
     }
 
     static void getDataFromMap(Environment environment, BObject visitor, BObject node, BMap<BString, Object> map,
-                               BMap<BString, Object> data) {
+                               BMap<BString, Object> data, CallbackHandler callbackHandler) {
         BMap<BString, Object> arguments = getArgumentsFromField(node);
         BString key = arguments.getStringValue(StringUtils.fromString(KEY));
         Object result = map.get(key);
-        getDataFromResult(environment, visitor, node, result, data);
+        getDataFromResult(environment, visitor, node, result, data, callbackHandler);
+    }
+
+    static void getDataFromArray(Environment environment, BObject visitor, BObject node, BArray result,
+                                 BMap<BString, Object> data, CallbackHandler callbackHandler) {
+        if (isScalarType(result.getElementType())) {
+            data.put(node.getStringValue(NAME_FIELD), result);
+        } else {
+            BArray resultArray = ValueCreator.createArrayValue(getDataRecordArrayType());
+            for (int i = 0; i < result.size(); i++) {
+                Object resultElement = result.get(i);
+                BMap<BString, Object> subData = createDataRecord();
+                getDataFromResult(environment, visitor, node, resultElement, subData, callbackHandler);
+                resultArray.append(subData.get(node.getStringValue(NAME_FIELD)));
+            }
+            data.put(node.getStringValue(NAME_FIELD), resultArray);
+        }
+    }
+
+    private static void getDataFromTable(Environment environment, BObject visitor, BObject node, BTable table,
+                                         BMap<BString, Object> data, CallbackHandler callbackHandler) {
+        Object[] valueArray = table.values().toArray();
+        ArrayType arrayType = TypeCreator.createArrayType(((BValue) valueArray[0]).getType());
+        BArray valueBArray = ValueCreator.createArrayValue(valueArray, arrayType);
+        getDataFromResult(environment, visitor, node, valueBArray, data, callbackHandler);
     }
 
     static void processFragmentNodes(Environment environment, BObject visitor, BObject node,
-                                     BMap<BString, Object> record, BMap<BString, Object> data) {
+                                     BMap<BString, Object> record, BMap<BString, Object> data,
+                                     CallbackHandler callbackHandler) {
         BArray selections = node.getArrayValue(SELECTIONS_FIELD);
         for (int i = 0; i < selections.size(); i++) {
             BMap<BString, Object> selection = (BMap<BString, Object>) selections.get(i);
             boolean isFragment = selection.getBooleanValue(IS_FRAGMENT_FIELD);
             BObject subNode = selection.getObjectValue(NODE_FIELD);
             if (isFragment) {
-                processFragmentNodes(environment, visitor, subNode, record, data);
+                processFragmentNodes(environment, visitor, subNode, record, data, callbackHandler);
             } else {
                 BString fieldName = subNode.getStringValue(NAME_FIELD);
                 Object fieldValue = record.get(fieldName);
-                getDataFromResult(environment, visitor, subNode, fieldValue, data);
+                getDataFromResult(environment, visitor, subNode, fieldValue, data, callbackHandler);
             }
         }
     }
 
     private static void executeResourceForFragmentNodes(Environment environment, BObject service, BObject visitor,
-                                                        BObject node, BMap<BString, Object> data, List<String> paths) {
+                                                        BObject node, BMap<BString, Object> data, List<String> paths,
+                                                        CallbackHandler callbackHandler) {
         BArray selections = node.getArrayValue(SELECTIONS_FIELD);
-        BMap<BString, Object> subData = createDataRecord();
         for (int i = 0; i < selections.size(); i++) {
             BMap<BString, Object> selection = (BMap<BString, Object>) selections.get(i);
             boolean isFragment = selection.getBooleanValue(IS_FRAGMENT_FIELD);
             BObject subNode = selection.getObjectValue(NODE_FIELD);
             if (isFragment) {
-                getDataFromService(environment, service, visitor, subNode, subData, paths);
-                data.merge(subData.getMapValue(subNode.getStringValue(NAME_FIELD)), false);
+                executeResourceForFragmentNodes(environment, service, visitor, subNode, data, paths, callbackHandler);
             } else {
-                List<String> updatedPaths = copyAndUpdateResourcePathsList(paths, subNode);
-                executeResource(environment, service, visitor, subNode, subData, updatedPaths);
-                data.put(subNode.getStringValue(NAME_FIELD), subData.get(subNode.getStringValue(NAME_FIELD)));
+                executeResourceWithPath(environment, visitor, subNode, service, data, paths, callbackHandler);
             }
         }
     }
 
-    private static void getDataFromTable(Environment environment, BObject visitor, BObject node, BTable table,
-                                         BMap<BString, Object> data) {
-        Object[] valueArray = table.values().toArray();
-        ArrayType arrayType = TypeCreator.createArrayType(((BValue) valueArray[0]).getType());
-        BArray valueBArray = ValueCreator.createArrayValue(valueArray, arrayType);
-        getDataFromArray(environment, visitor, node, valueBArray, data);
-    }
-
-    private void appendErrorToVisitor(BError bError) {
-        BArray errors = this.visitor.getArrayValue(ERRORS_FIELD);
-        errors.append(getErrorDetailRecord(bError, this.node));
+    private static void executeResourceWithPath(Environment environment, BObject visitor, BObject node, BObject service,
+                                                BMap<BString, Object> data, List<String> paths,
+                                                CallbackHandler callbackHandler) {
+        List<String> updatedPaths = copyAndUpdateResourcePathsList(paths, node);
+        executeResource(environment, service, visitor, node, data, updatedPaths, callbackHandler);
     }
 
     private static ArrayType getDataRecordArrayType() {